refactor: separate 42 oauth code to views_42.py
